Initial commit: Project files and Basic Controls
This would be the initial commit which adds the
project files and implements the basic controls
such as swipe up to jump, swipe left and right
to dash and swipe down to bring the player down.
I also added desktop controls where they can
control the character using WASD keys.

Signed-off-by: Christopher Quinto <cquinto.primary@gmail.com>