Add Tutorial Part 1
I added new assets such as "hand.png" for the swipe
gesture, "red.png" that would roll on the ground,
and "spike.png" that would fall from the ceiling.

I modified the InputHandler.js so that I can
enable and disable the player controls at certain
parts of the tutorial.

I added the TutorialManager.js file which includes
all the tutorial sequences.

I added the new configurations on the Game.js
which is the game orchestrator.

Signed-off-by: Christopher Quinto <cquinto.primary@gmail.com>